Move Internals::byte_array into crypto.cpp
